♻️ deleteEntityが存在する場合にのみ実行するように修正。

+++ vite/src/components/entity/form/EntityModal.tsx
@@ -79,7 +79,7 @@ const EntityModal = (props: {
               <Button
                 type="submit"
                 onClick={() => {
-                  dispatch(deleteEntity(props.entity!.id));
+                  deleteEntity && dispatch(deleteEntity(props.entity!.id));
                 }}
               >
                 {submitText}
